requirements.txt
discord.py==0.16.12
apiai
numpy
praw==5.1.0
wikia==1.4.3
humanize
psutil
pillow
pytz
requests
wikipedia
wiktionaryparser
psycopg2
